backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0 